Set up the Control repo for C++ development
Removed previous python code from ths branch to allow for a cleaner C++
development. Added the .h and .cpp files for some planned Classes. Added
prj/README.md to explain the C++ naming conventions that we will be
sticking to to ensure consistency through the team. Added clang-format
to this branch, as well as pre-commit to run the clang-format as an
interupt when calling git commit so that everyone pushes code that gets
formatted the same way.

Created by:     Dawson van Vlaanderen
From team:      Control Software