Added commented out structs to Message class
I added some (commented out) code for a way to use structs instead of
vectors for the instructions of a message. Basically, there are
different structs for the different message types we want to send.
When being constructed, the Message class uses the MessagePayload type
(which uses std::variant) to choose the type of message we want.

This allows us to specify many different types of Messages without
creating children classes of Message, and increases readability,
documentation, andability to change messages more easily in the future.