Updated main and added mission_control.h
Added checks to main to only compile the parts of the code pertaining to
a specific build (ie only for rover or only for base). Moved
the #defines that are needed for these configurations to
"mission_control.h" (thought a file that contains a list of
configurations should have a fun name like this lol).

Created by:     Dawson van Vlaanderen
From team:      Control Software