Updated submodules to contain dependancies
Libraries that are used in the project have been added as submodules to
the repository. This allows for easier management of the libraries and
allows others to clone the repository without needing to download the
libraries separately. The libraries are located in the tools directory.
The Makefile has been updated to work with these library locations.

With this change, I fixed the motor initialization based on my testing
in another file.

Created by:     Dawson van Vlaanderen
From team:      Control Software