Added infrastructure for google tests
Added the cmake file and initial files for the gtests. to run, tests
only need to be added. Need to work on splitting up base class so that
the file containing functions for changing vars ins not dependant on SDL
or other majror libraries.

Created by:     Dawson van Vlaanderen
From team:      Control Softeware